PkgGraph

Documentation for PkgGraph.

PkgGraph.PkgGraph

PkgGraph.style

PkgGraph.create

PkgGraph.depgraph

PkgGraph.deps_as_DOT

PkgGraph.open

PkgGraph.set_base_url

PkgGraph.to_DOT_str

PkgGraph.url

PkgGraph.PkgGraph — Module

Visualize the dependency graph of a Julia package.

Use PkgGraph.open to view the graph in the browser, or PkgGraph.create to generate an image locally.

PkgGraph.style — Constant

style

A list of strings used in constructing the "digraph" string in to_DOT_str. They are insterted as lines just before the graph edge lines.

To use the default Graphviz style, call:

empty!(PkgGraph.style)

See Graphviz Attributes for more info on how dot graphs can be styled.

PkgGraph.create — Method

create(pkgname; dir = tempdir(), fmt = :png)

Render the dependency graph of the given package as an image in dir, and open it with your default image viewer. Uses the external program 'dot' (https://graphviz.org), which must be available on PATH.

fmt is an output file format supported by dot, such as svg or png.

PkgGraph.depgraph — Method

deps = depgraph(pkgname)

Build a graph of the dependencies of the given package, using the active project's Manifest file.

The returned deps object is a flat list of "PkgA" => "PkgB" dependency pairs.

PkgGraph.deps_as_DOT — Method

deps_as_DOT(pkgname)

Create the dependency graph of pkgname and render it as a Graphviz DOT string.

Example output (truncated), for "Unitful":

digraph {
    node [fontname = "sans-serif"]
    edge [arrowsize = 0.88]
    Unitful -> ConstructionBase
    ConstructionBase -> LinearAlgebra
    LinearAlgebra -> Libdl
    ⋮
    Unitful -> Random
    Random -> SHA
    Random -> Serialization
}

For more info, see depgraph and to_DOT_str.

PkgGraph.open — Method

open(pkgname)

Open the browser to an image of pkgname's dependency graph.
The given package must be installed in the currently active project.

To render the dependency graph using a local Graphviz dot installation (instead of an online Graphviz renderer), use create.

For more info, see depgraph_url.

PkgGraph.set_base_url — Method

set_base_url(new)

Set the rendering website that will be used by open and depgraph_url to the given new url.

See PkgGraph.base_urls for some options.

PkgGraph.to_DOT_str — Method

to_DOT_str(edges; indent = 4)

Build a string that represents the given directed graph in the Graphviz DOT format.

Example:

julia> empty!(PkgGraph.style);

julia> edges = [:A => :B, "yes" => "no"];

julia> PkgGraph.to_DOT_str(edges, indent = 2) |> println
digraph {
  A -> B
  yes -> no
}

PkgGraph.url — Method

depgraph_url(pkgname)

Create a URL at which the dependency graph of pkgname is rendered as an image, using an online Graphviz rendering service.

How it works

The dependency graph of pkgname is created locally, and converted to a string in the Graphviz DOT format (see deps_as_DOT). This string is URL-encoded, and appended to a partly-complete URL, which is by default the first entry in the PkgGraph.base_urls list. To use a different rendering website, use set_base_url.