Internals

Creating a dependency graph

PkgGraph.depgraph — Function

deps = depgraph(pkgname)

Build a graph of the dependencies of the given package, using the active project's Manifest file.

The returned deps object is a flat list of "PkgA" => "PkgB" dependency pairs.

DOT strings

PkgGraph.deps_as_DOT — Function

deps_as_DOT(pkgname)

Create the dependency graph of pkgname and render it as a Graphviz DOT string.

Example output (truncated), for "Unitful":

digraph {
    node [fontname = "sans-serif"]
    edge [arrowsize = 0.88]
    Unitful -> ConstructionBase
    ConstructionBase -> LinearAlgebra
    LinearAlgebra -> Libdl
    ⋮
    Unitful -> Random
    Random -> SHA
    Random -> Serialization
}

For more info, see depgraph and to_DOT_str.

PkgGraph.to_DOT_str — Function

to_DOT_str(edges; indent = 4)

Build a string that represents the given directed graph in the Graphviz DOT format.

Example:

julia> empty!(PkgGraph.style);

julia> edges = [:A => :B, "yes" => "no"];

julia> PkgGraph.to_DOT_str(edges, indent = 2) |> println
digraph {
  A -> B
  yes -> no
}

PkgGraph.style — Constant

style

A list of strings used in constructing the "digraph" string in to_DOT_str. They are insterted as lines just before the graph edge lines.

To use the default Graphviz style, call:

empty!(PkgGraph.style)

See Graphviz Attributes for more info on how dot graphs can be styled.

Online rendering

PkgGraph.url — Function

url(pkgname)

Create a URL at which the dependency graph of pkgname is rendered as an image, using an online Graphviz rendering service.

How it works

The dependency graph of pkgname is created locally, and converted to a string in the Graphviz DOT format (see deps_as_DOT). This string is URL-encoded, and appended to a partly-complete URL, which is by default the first entry in the PkgGraph.base_urls list. To use a different rendering website, use set_base_url.