End-user

PkgGraph.PkgGraph — Module

Visualize the dependency graph of a Julia package.

Use depgraph_web to view the graph in the browser, or depgraph_image to generate an image locally.

Online

PkgGraph.depgraph_web — Function

depgraph_web(pkgname, base_url = first(webapps); kw...)

Open the browser to an image of pkgname's dependency graph.

See url for more on base_url. Keyword arguments are passed on to depgraph_as_dotstr

Local

PkgGraph.depgraph_image — Function

depgraph_image(
    pkgname;
    dir    = tempdir(),
    fmt    = :png,
    bg     = bg(fmt),
    open   = true,
    post   = true,
    kw...
)

Render the dependency graph of the given package as an image in dir, and open it with your default image viewer. Uses the external program 'dot' (see graphviz.org), which must be available on PATH.

bg: background colour for the image. Default is :white if fmt = :png, and :transparent otherwise.

fmt is an output file format supported by dot, such as :svg or :png.
If fmt is :svg, and post is true (default), the generated SVG file is post-processed, to add light and dark-mode CSS.

To only create the image, without automatically opening it, pass open = false.

Other keyword arguments are passed on to depgraph_as_dotstr