Internal

PkgGraph.Internals — Module

Namespace for the non-end-user functions in PkgGraph.

For ease of experimentation, you can import these with

using PkgGraph.Internals

(They are also imported in the main module, so they can be accessed as PkgGraph.depgraph, e.g).

Creating a dependency graph

PkgGraph.Internals.depgraph — Function

depgraph(pkgname)

Build a directed graph of the dependencies of the given package, using the active project's Manifest file.

The returned deps object is a flat list of "PkgA" => "PkgB" dependency pairs.

Example:

julia> using PkgGraph.Internals

julia> depgraph(:Test)
8-element Vector{Any}:
             "Test" => "InteractiveUtils"
 "InteractiveUtils" => "Markdown"
         "Markdown" => "Base64"
             "Test" => "Logging"
             "Test" => "Random"
           "Random" => "SHA"
           "Random" => "Serialization"
             "Test" => "Serialization"

PkgGraph.Internals.packages_in_active_manifest — Function

packages_in_active_manifest()

Parsed contents of the 'dependencies' part of the active project's Manifest.toml.

DOT strings

PkgGraph.Internals.deps_as_dot — Function

deps_as_dot(pkgname; kw...)

Create the dependency graph of pkgname and render it as a Graphviz DOT string.

See to_dot_str for keyword arguments.

Example:

julia> PkgGraph.deps_as_dot(:Test) |> println
digraph {
    bgcolor = "transparent"
    node [fontname = "sans-serif", style = "filled", fillcolor = "white"]
    edge [arrowsize = 0.88]
    Test -> InteractiveUtils
    InteractiveUtils -> Markdown
    Markdown -> Base64
    Test -> Logging
    Test -> Random
    Random -> SHA
    Random -> Serialization
    Test -> Serialization
}

PkgGraph.Internals.to_dot_str — Function

to_dot_str(edges; indent = 4, emptymsg = nothing, style = default_style)

Build a string that represents the given directed graph in the Graphviz DOT format ↗.

If there are no edges, a single node with emptymsg is created.

style is a list of strings, inserted as lines in the output (just before the graph edge lines). To use Graphviz's default style, pass style = []. For more on how dot-graphs can be styled, see 'Graphviz Attributes' ↗

Example:

julia> using PkgGraph.Internals

julia> edges = [:A => :B, "yes" => "no"];

julia> to_dot_str(edges, indent=2, style=[]) |> println
digraph {
  A -> B
  yes -> no
}

PkgGraph.Internals.default_style — Constant

The default style used by to_dot_str:

bgcolor = "transparent"
node [fontname = "sans-serif", style = "filled", fillcolor = "white"]
edge [arrowsize = 0.88]

Online rendering

PkgGraph.Internals.url — Function

url(pkgname, base_url = first(rendering_websites); kw...)

Create a URL at which the dependency graph of pkgname is rendered as an image, using an online Graphviz rendering service.

Keyword arguments are passed on to deps_as_dot and to_dot_str.

How it works

The dependency graph of pkgname is created locally, and converted to a string in the Graphviz DOT format. This string is URL-encoded, and appended to a partly-complete base_url.

Example:

julia> base_url = PkgGraph.rendering_websites[2]
"http://magjac.com/graphviz-visual-editor/?dot="

julia> PkgGraph.url(:TOML, base_url, style=[], indent=0)
"http://magjac.com/graphviz-visual-editor/?dot=digraph%20%7B%0ATOML%20-%3E%20Dates%0ADates%20-%3E%20Printf%0APrintf%20-%3E%20Unicode%0A%7D%0A"