Internals

Creating a dependency graph

PkgGraph.DepGraph.depgraph — Function

depgraph(pkgname; jll = true, stdlib = true)

Build a directed graph of the dependencies of the given package, using the active project's Manifest file.

The returned deps object is a flat list of "PkgA" => "PkgB" dependency pairs.

Binary JLL dependencies and packages in the standard library can be filtered out from the result by setting jll and stdlib to false.

Example:

julia> using PkgGraph

julia> PkgGraph.depgraph(:Test)
8-element Vector{Pair{String, String}}:
             "Test" => "InteractiveUtils"
 "InteractiveUtils" => "Markdown"
         "Markdown" => "Base64"
             "Test" => "Logging"
             "Test" => "Random"
           "Random" => "SHA"
           "Random" => "Serialization"
             "Test" => "Serialization"

PkgGraph.DepGraph.is_jll — Function

is_jll(pkg)::Bool

PkgGraph.DepGraph.is_in_stdlib — Function

is_in_stdlib(pkg)::Bool

DOT strings

PkgGraph.DotString.to_dot_str — Function

to_dot_str(
    edges;
    dark     = false,
    bg       = "transparent",
    style    = default_style(),
    indent   = 4,
    emptymsg = nothing,
)

Build a string that represents the given directed graph in the Graphviz DOT format ↗.

Keyword arguments

dark

Default is false i.e. light-mode: black lines and black text with white backgrounds. For dark = true, it's white lines and white text, with black backgrounds. Note that locally-generated SVGs get both colour-schemes simultaneously, so this option is irrelevant for them.

bg

Background colour for the image.
Default is "transparent".
"white" (in combination with dark = false) might be a sensible value when you are creating a PNG but do not know on what background it will be seen. (A light-mode PNG with transparent background looks bad on a dark background).

style

A list of strings, inserted as lines in the output (after the lines generated for dark and bg, and just before the graph edge lines). To use Graphviz's default style, pass style = []. For the default see default_style. For more on how dot-graphs can be styled, see Styling Graphviz output.

indent

The number of spaces to indent each line in the "digraph" block with.

emptymsg

If there are no edges, a single node with emptymsg is created. If emptymsg is nothing (default), no nodes are created, and the image rendered from the DOT-string will be empty.

Example:

julia> edges = [:A => :B, "yes" => "no"];

julia> style = ["node [color=\"red\"]"];

julia> using PkgGraph

julia> PkgGraph.to_dot_str(edges; style, bg=:blue, indent=2) |> println
digraph {
  bgcolor = "blue"
  node [fillcolor="white", fontcolor="black", color="black"]
  edge [color="black"]
  node [color="red"]
  A -> B
  yes -> no
}

julia> emptymsg="(empty graph)";

julia> PkgGraph.to_dot_str([]; emptymsg, dark=true, style=[]) |> println
digraph {
    bgcolor = "transparent"
    node [fillcolor="black", fontcolor="white", color="white"]
    edge [color="white"]
    onlynode [label=" (empty graph) ", shape="plaintext"]
}

PkgGraph.DotString.default_style — Function

default_style()

The default style used by to_dot_str:

node [fontname="sans-serif", style="filled"]
edge [arrowsize=0.88]

Post-processing of SVG files

PkgGraph.SVG.add_darkmode — Function

add_darkmode(path, out = path)

Edit an SVG file generated by Graphviz.

Remove repeated styling on individual elements, and replace it by a global stylesheet, with both light and dark modes. (This is achieved through a CSS @media query for prefers-color-scheme).

Online rendering

PkgGraph.url — Function

url(dotstr, base = first(webapps))

Create a URL at which the given dot-string is rendered as an image, using an online Graphviz rendering service.

The dot-string is URL-encoded, and appended to a partly complete base URL (see webapps)

Example:

julia> base = PkgGraph.webapps[2];

julia> PkgGraph.url("digraph {Here->There}", base)
"http://magjac.com/graphviz-visual-editor/?dot=digraph%20%7BHere-%3EThere%7D"

PkgGraph.webapps — Constant

A list of websites that can render Graphviz dot-formatted strings. Used by url.

Note that these are 'base URLs', to which url-encoded dot-strings can be directly appended.

Default contents:

https://dreampuf.github.io/GraphvizOnline/#

http://magjac.com/graphviz-visual-editor/?dot=

https://edotor.net/?engine=dot#

Graphs.jl conversion

PkgGraph.DepGraph.as_graphsjl_input — Function

as_graphsjl_input(edges)

Obtain the outputs of, respectively

vertices(edges)

node_index(edges)

adjacency_matrix(edges)

..but without unnecessary repeat computation.

Useful when converting the output of depgraph to a Graphs.jl graph;
See the example script in Working with Graphs.jl.

PkgGraph.DepGraph.vertices — Function

vertices(edges)

Extract the unique nodes from the given list of edges.

Example:

julia> using PkgGraph

julia> edges = PkgGraph.depgraph(:Test);

julia> PkgGraph.vertices(edges)
8-element Vector{String}:
 "Test"
 "InteractiveUtils"
 "Markdown"
 "Random"
 "Base64"
 "Logging"
 "SHA"
 "Serialization"

PkgGraph.DepGraph.node_index — Function

node_index(edges)
node_index(vertices)

Create a function that returns the index of a given vertex.

This is useful because Graphs.jl requires vertices to be integers.

Example:

julia> using PkgGraph

julia> edges = ["A"=>"B", "B"=>"C"];

julia> node = PkgGraph.node_index(edges);

julia> node("C")
3

PkgGraph.DepGraph.adjacency_matrix — Function

adjacency_matrix(edges)

A square bitmatrix A that is 0 everywhere except at A[i,j] when there is a connection from the node with index i to the node with index j.