Usage

PkgGraph.PkgGraph — Module

Visualize the dependency graph of a Julia package.

Use PkgGraph.open to view the graph in the browser, or PkgGraph.create to generate an image locally.

PkgGraph.open — Function

open(pkgname)

Open the browser to an image of pkgname's dependency graph.
The given package must be installed in the currently active project.

To render the dependency graph using a local Graphviz dot installation (instead of an online Graphviz renderer), use create.

For more info, see url.

PkgGraph.create — Function

create(pkgname; dir = tempdir(), fmt = :png)

Render the dependency graph of the given package as an image in dir, and open it with your default image viewer. Uses the external program 'dot' (graphviz.org), which must be available on PATH.

fmt is an output file format supported by dot, such as svg or png.

PkgGraph.set_base_url — Function

set_base_url(new)

Set the rendering website that will be used by open and url to the given new url.

See PkgGraph.base_urls for some options.