PkgGraph

Documentation for PkgGraph.

PkgGraph.create_depgraph

PkgGraph.depgraph

PkgGraph.depgraph_local

PkgGraph.depgraph_url

PkgGraph.deps_as_DOT

PkgGraph.set_rendering_url

PkgGraph.to_DOT_str

PkgGraph.create_depgraph — Method

deps = create_depgraph(pkgname)

Build a graph of the dependencies of the given package, using the active project's Manifest file.

The returned deps object is a flat list of "PkgA" => "PkgB" dependency pairs.

PkgGraph.depgraph — Method

depgraph(pkgname)

Open the browser to an image of pkgname's dependency graph.
The given package must be installed in the currently active project.

To render the dependency graph using a local Graphviz dot installation (instead of an online Graphviz renderer), use depgraph_local.

For more info, see depgraph_url.

PkgGraph.depgraph_local — Function

depgraph_local(pkgname, dir = tempdir(); fmt = :png)

Render the dependency graph of the given package as an image in dir, and open it with your default image viewer. Uses the external program 'dot' (https://graphviz.org), which must be available on PATH.

fmt is an output file format supported by dot, such as svg or png.

PkgGraph.depgraph_url — Method

depgraph_url(pkgname)

Create a URL at which the dependency graph of pkgname is rendered as an image, using an online Graphviz rendering service.

How it works

The dependency graph of pkgname is created locally, and converted to a string in the Graphviz DOT format (see deps_as_DOT). This string is URL-encoded, and appended to a partly-complete URL, which is by default the first entry in the PkgGraph.rendering_urls list. To use a different rendering website, use set_rendering_url.

PkgGraph.deps_as_DOT — Method

deps_as_DOT(pkgname)

Create the dependency graph of pkgname and render it as a Graphviz DOT string.

Example output (truncated), for "Unitful":

digraph {
    Unitful -> ConstructionBase
    ConstructionBase -> LinearAlgebra
    LinearAlgebra -> Libdl
    ⋮
    Unitful -> Random
    Random -> SHA
    Random -> Serialization
}

For more info, see create_depgraph and to_DOT_str.

PkgGraph.set_rendering_url — Method

set_rendering_url(new)

Set the base url that will be used by depgraph and depgraph_url to the given new url.

See PkgGraph.rendering_urls for some options.

PkgGraph.to_DOT_str — Method

to_DOT_str(edges)

Build a string that represents the given directed graph in the Graphviz DOT format (https://graphviz.org/doc/info/lang.html).

Example:

julia> using PkgGraph: to_DOT_str

julia> edges = [:A => :B, "yes" => "no"];

julia> to_DOT_str(edges) |> println
digraph {
    A -> B
    yes -> no
}