End-user

PkgGraphs.PkgGraphs — Module

Visualize the dependency graph of a Julia package.

Use PkgGraphs.open to view the graph in the browser, or PkgGraphs.create to generate an image locally.

(Note that these functions are not exported).

See PkgGraphs.Internals for more functions.

PkgGraphs.open — Function

open(pkgname, base_url = first(webapps))

Open the browser to an image of pkgname's dependency graph.

The given package must be installed in the currently active project.

See also webapps and Internals.url.

PkgGraphs.create — Function

create(pkgname, dir = tempdir(); fmt = :png)

Render the dependency graph of the given package as an image in dir, and open it with your default image viewer. Uses the external program 'dot' (see graphviz.org), which must be available on PATH.

fmt is an output file format supported by dot, such as svg or png.

The given package must be installed in the currently active project.

PkgGraphs.webapps — Constant

A list of websites that can render Graphviz dot-formatted strings. Used by Internals.url.

Note that these are 'base URLs', to which url-encoded dot-strings can be directly appended.

Default contents:

https://dreampuf.github.io/GraphvizOnline/#

http://magjac.com/graphviz-visual-editor/?dot=

https://edotor.net/?engine=dot#