Internal

PkgGraph.Internals — Module

Namespace for the non-end-user functions in PkgGraph.

For ease of experimentation, you can import these with

using PkgGraph.Internals

They are also imported in the main module, so they can be accessed as PkgGraph.depgraph, e.g.

Creating a dependency graph

PkgGraph.Internals.depgraph — Function

depgraph(pkgname; include_jll = true, include_stdlib = true)

Build a directed graph of the dependencies of the given package, using the active project's Manifest file.

The returned deps object is a flat list of "PkgA" => "PkgB" dependency pairs.

Binary JLL dependencies and packages in the standard library can be filtered out from the result by setting include_jll and include_stdlib to false.

Example:

julia> using PkgGraph.Internals

julia> depgraph(:Test)
8-element Vector{Pair{String, String}}:
             "Test" => "InteractiveUtils"
 "InteractiveUtils" => "Markdown"
         "Markdown" => "Base64"
             "Test" => "Logging"
             "Test" => "Random"
           "Random" => "SHA"
           "Random" => "Serialization"
             "Test" => "Serialization"

PkgGraph.Internals.vertices — Function

vertices(edges)

Extract the unique nodes from the given list of edges.

Useful when converting the output of depgraph to a Graphs.jl graph. (See the example script in The Graphs.jl ecosystem).

Example:

julia> using PkgGraph.Internals

julia> edges = depgraph(:Test);

julia> vertices(edges)
8-element Vector{String}:
 "Test"
 "InteractiveUtils"
 "Markdown"
 "Random"
 "Base64"
 "Logging"
 "SHA"
 "Serialization"

PkgGraph.Internals.packages_in_active_manifest — Function

packages_in_active_manifest()

Read and parse the Manifest.toml of the active project, and return its 'deps' table (as a dictionary indexed by package names).

Every entry in this dictionary is a list. This is for when multiple packages would share the same name.

Example:

julia> using PkgGraph.Internals

julia> packages = packages_in_active_manifest();

julia> only(packages["PkgGraph"])
Dict{String, Any} with 4 entries:
  "deps"    => ["DefaultApplication", "TOML", "URIs"]
  "uuid"    => "f9c1b9e4-72e8-4a14-ade5-14f45fc35f11"
  "version" => "0.1.0"
  "path"    => "C:\Users\tfiers\.julia\dev\PkgGraph"

DOT strings

PkgGraph.Internals.to_dot_str — Function

to_dot_str(edges; style = default_style, indent = 4, emptymsg = nothing)

Build a string that represents the given directed graph in the Graphviz DOT format ↗.

style is a list of strings, inserted as lines in the output (just before the graph edge lines). To use Graphviz's default style, pass style = []. For more on how dot-graphs can be styled, see Styling Graphviz output.

indent is the number of spaces to indent each line in the "digraph" block with.

If there are no edges, a single node with emptymsg is created. If emptymsg is nothing (default), no nodes are created, and the image rendered from the DOT-string will be empty.

Example:

julia> edges = [:A => :B, "yes" => "no"];

julia> style = ["node [color = \"red\"]"];

julia> using PkgGraph.Internals

julia> to_dot_str(edges; style, indent = 2) |> println
digraph {
  node [color = "red"]
  A -> B
  yes -> no
}

julia> to_dot_str([], style=[], emptymsg="(empty graph)") |> println
digraph {
    onlynode [label = " (empty graph) ", shape = "plaintext"]
}

See also default_style.

PkgGraph.Internals.default_style — Constant

The default style used by to_dot_str:

bgcolor = "transparent"
node [fontname = "sans-serif", style = "filled", fillcolor = "white"]
edge [arrowsize = 0.88]

Online rendering

PkgGraph.Internals.url — Function

url(base, dotstr)

Create a URL at which the given dot-string is rendered as an image, using an online Graphviz rendering service.

The dot-string is URL-encoded, and appended to a partly complete base URL (see webapps)

Example:

julia> base = PkgGraph.webapps[2];

julia> PkgGraph.url(base, "digraph {Here->There}")
"http://magjac.com/graphviz-visual-editor/?dot=digraph%20%7BHere-%3EThere%7D"

PkgGraph.Internals.webapps — Constant

A list of websites that can render Graphviz dot-formatted strings. Used by Internals.url.

Note that these are 'base URLs', to which url-encoded dot-strings can be directly appended.

Default contents:

https://dreampuf.github.io/GraphvizOnline/#

http://magjac.com/graphviz-visual-editor/?dot=

https://edotor.net/?engine=dot#